- maven format - removed most nms - Erumeldor can use it
