--- messages.yml
@@ -1,29 +1,0 @@
-#player interactions
-usage: "To create a quantum circuit, use /qc <circuit>; and click on a sender and then a receiver with redstone."
-available_circuits: "Available circuits:"
-cancelled: "Your pending circuit has been removed!"
-no_pending_circuit: "No pending circuit to remove."
-circuit_created: "Quantum circuit created!"
-no_sender: "You need to setup a sender and receiver first!"
-no_receivers: "You need to setup at least one receiver first!"
-no_pending_action: "No pending action to finish."
-invalid_delay: "Invalid delay time (max:%maxdelay%s), assuming no delay"
-circuit_ready: "Circuit ready to be created! (%circuit% %delay%s delay)"
-circuit_changed: "Circuit type switched to: (%circuit% %delay%s delay)"
-no_permission: "You don't have permission to create the %circuit% circuit"
-invalid_circuit: "Invalid circuit specified. '/qc' for usage."
-
-#log messages
-console_not_allowed: "You can't run QC commands from the console!"
-found_old_file: "Found %file%, attempting to convert..."
-loading: "Loading %file%..."
-loading_not_found: "%file% not found, will be created with the next save."
-loading_no_circuits: "No circuits found in %file%"
-saving: "Saving %file%..."
-saved: "%file% Saved!"
-save_failed: "%world% could not be saved!"
-save_failed_file: "Unable to save %file%!"
-receiver_removed: "Removed a %world% circuit's receiver; %block% is not a valid receiver."
-circuit_removed_invalid: "Removed a %world% circuit; %block% is not a valid sender."
-circuit_removed_no_receivers: "Removed a '%world%' circuit: no valid receivers."
-
