+++ src/main/resources/plugin.yml
@@ -1,0 +1,46 @@
+name: QuantumConnectors
+main: com.ne0nx3r0.quantum.QuantumConnectors
+version: '1.10.2'
+database: false
+startup: postworld
+description: Wireless redstone connections
+commands:
+   qc:
+      description: Create a quantum connector
+permissions:
+    QuantumConnectors.*:
+      description: Allows all permissions
+      children:
+        QuantumConnectors.create.*: true
+    QuantumConnectors.create.*:
+      description: Allows creation of circuits
+      children:
+        QuantumConnectors.create.quantum: true
+        QuantumConnectors.create.on: true
+        QuantumConnectors.create.off: true
+        QuantumConnectors.create.toggle: true
+        QuantumConnectors.create.reverse: true
+        QuantumConnectors.create.random: true 
+        QuantumConnectors.update: true 
+        #purposely not including spawn here, so it can't accidentally be granted
+    QuantumConnectors.create.quantum:
+      description: Allows creation of "Quantum" Circuit
+      default: op
+    QuantumConnectors.create.on:
+      description: Allows creation of "On" circuit
+      default: op
+    QuantumConnectors.create.off:
+      description: Allows creation of "Off" circuit
+      default: op
+    QuantumConnectors.create.toggle:
+      description: Allows creation of "Toggle" circuit
+      default: op
+    QuantumConnectors.create.reverse:
+      description: Allows creation of "Reverse" circuit
+      default: op
+    QuantumConnectors.create.random:
+      description: Allows creation of "Random" circuit
+      default: op
+    QuantumConnectors.create.spawn:
+      description: Allows creation of "Spawn" circuit
+      default: op
